adding static main to original benchmarks for running spf, also pushing in the changes on the SymbolicSequenceListener

+++ src/main/gov/nasa/jpf/symbc/sequences/SymbolicSequenceListener.java
@@ -469,7 +469,7 @@ public class SymbolicSequenceListener extends PropertyListenerAdapter implements
                         else
                             solution = solution + "true";
                     } else
-                        solution = solution + ((IntegerExpression) e).solution();
+                        solution = solution + (int)((IntegerExpression) e).solution();
                 } else if (e instanceof RealExpression)
                     solution = solution + ((RealExpression) e).solution();
                 else
